% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkConstruction.R
\name{blendPT3d}
\alias{blendPT3d}
\title{Blend 3D Points into 3D Linestrings}
\usage{
blendPT3d(x, y, restriction = NULL)
}
\arguments{
\item{x}{An \code{sf} object with 3D \code{POINT} geometry. The points to be blended into the linestrings.}

\item{y}{An \code{sf} object with 3D \code{LINESTRING} geometry. The linestrings to receive the projected points as new vertices.}

\item{restriction}{A \code{character} value specifying a field in \code{y} that indicates whether each linestring can be blended. The field must contain only \code{integer} values 0 (not restricted) or 1 (restricted). If \code{NULL} (default), no restrictions apply.}
}
\value{
An \code{sf} object with 3D \code{LINESTRING} geometry, where projected points from \code{x} are incorporated as additional vertices in \code{y}, respecting any \code{restriction} field.
}
\description{
Projects each 3D point onto its nearest 3D linestring and incorporates the projected points as new vertices, with an optional restriction to prohibit blending into specific linestrings.
}
\details{
For each 3D point in \code{x}, the function identifies its nearest location (via 3D orthogonal projection) on the linestrings in \code{y} and adds these projected points as new vertices. If \code{restriction} is specified, linestrings with a restriction value of 1 are not modified.
}
\examples{
\dontrun{
library(sf)
library(rgl)

# Create a 3D linestrings
y = st_sf(geometry = st_sfc(
  st_linestring(matrix(c(0,0,0, 1,2,1), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(1,2,1, 3,3,2), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(3,3,2, 5,2,3), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(5,2,3, 7,0,4), ncol=3, byrow=TRUE)))
)

y$restriction = c(1,0,0,1)

# Create some points near the line
x = st_sf(geometry = st_sfc(
  st_point(c(1,1,1)),
  st_point(c(4,2.5,3.5)),
  st_point(c(6,0.5,3.5))
)) \%>\% st_as_sf()

# Blend points into linestring
blended = blendPT3d(x, y) # Without restriction
blended.restriction = blendPT3d(x, y, "restriction") # With restriction

# Print result
st_coordinates(blended)
st_coordinates(blended.restriction)

}
}
