% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialQuery.R
\name{nearest3d}
\alias{nearest3d}
\title{Find the Nearest Feature in 3D Space}
\usage{
nearest3d(x, y)
}
\arguments{
\item{x}{An \code{sf} object with 3D \code{POINT} geometry, representing the query points.}

\item{y}{An \code{sf} object with 3D \code{POINT} or \code{LINESTRING} geometry, representing the target features.}
}
\value{
A \code{numeric} vector of length \code{nrow(x)}, where each value is the index of the nearest feature in \code{y} for the corresponding point in \code{x}.
}
\description{
Identifies the index of the nearest 3D point or linestring for each input 3D point.
}
\details{
The function finds the nearest feature in \code{y} for each 3D point in \code{x} using 3D Euclidean distance, supporting two scenarios:
\itemize{
  \item \code{POINT}-to-\code{POINT}: Identifies the index of the nearest point in \code{y} for each point in \code{x}.
  \item \code{POINT}-to-\code{LINESTRING}: Identifies the index of the nearest linestring in \code{y} for each point in \code{x}, based on the closest point on the linestring.
}
}
\examples{
\dontrun{
library(sf)

# Example: point-to-point
pt1 = st_sf(geometry = st_sfc(
  st_point(c(0,0,0)),
  st_point(c(1,0,0)),
  st_point(c(0,1,1))
))

pt2 = st_sf(geometry = st_sfc(
  st_point(c(0,0,1)),
  st_point(c(1,1,2)),
  st_point(c(0,1,4.5))
))

print(nearest3d(pt1, pt2))

# Example: point-to-line
pt = st_sf(geometry = st_sfc(
  st_point(c(0,0,1)),
  st_point(c(1,1,2)),
  st_point(c(0,1,4.5)) )
  )
lines = st_sf(geometry = st_sfc(
  st_linestring(matrix(c(0,0,0, 1,1,1), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(2,0,0, 2,2,2), ncol=3, byrow=TRUE))
))

print(nearest3d(pt, lines))
}

}
