% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkConstruction.R
\name{nearestNode3d.rst}
\alias{nearestNode3d.rst}
\title{Find the Nearest Network Node for Each 3D Point with Restriction}
\usage{
nearestNode3d.rst(net, pt, restriction)
}
\arguments{
\item{net}{A \code{list} containing \code{nodes} (3D \code{sf} \code{POINT} objects with \code{character} \code{nID} column) and \code{edges} (3D \code{sf} \code{LINESTRING} objects), typically from \code{\link[GISnetwork3D]{sf_to_net3d}}.}

\item{pt}{An \code{sf} object with 3D \code{POINT} geometry, representing the input points for which to find the nearest network node.}

\item{restriction}{A \code{character} value specifying a field in \code{net$edges} with \code{integer} values (0 = not restricted, 1 = restricted).}
}
\value{
An \code{sf} object with 3D \code{POINT} geometry, identical to \code{pt}, with an additional column \code{nID} specifying the index of the nearest non-restricted network node for each point.
}
\description{
Identifies the nearest 3D network node for each input 3D point, excluding nodes connected to restricted edges, and attaches its index (\code{nID}).
}
\details{
For each 3D point in \code{pt}, the function identifies the closest 3D node in \code{net$nodes} based on 3D Euclidean distance, excluding nodes connected to edges where the \code{restriction} field is 1. The \code{restriction} parameter ensures that only nodes from non-restricted edges (value 0) are considered. The corresponding \code{nID} is appended as a new column to the output.
}
\examples{
\dontrun{
library(sf)

pts = st_sf(geometry = st_sfc(
  st_point(c(1,1,2)),
  st_point(c(6,6,3)),
  st_point(c(11, 11, 7))
))

# Create a line network
l <- st_sf(geometry = st_sfc(
  st_linestring(matrix(c(0,0,0, 2,2,2), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(2,2,2, 5,5,2), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(5,5,2, 9,9,5), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(9,9,5, 11,11,6), ncol=3, byrow=TRUE)),
  st_linestring(matrix(c(11,11,6, 13,13,6), ncol=3, byrow=TRUE))
))
l$restriction = c(0,1,1,0,0) # Assign restriction

# Create network
net <- sf_to_net3d.full(l, direction = rep(0, nrow(l)), CF = "Tobler") # Assumes bi-directional

# Find nearest nodes
pts.noRestriction = nearestNode3d(net$nodes, pts) # Without restriction
pts.restriction = nearestNode3d.rst(net, pts, "restriction") # With restriction
}

}
